// src/api/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        CameraAnalyticsSchema: {
            /** Total Cameras */
            total_cameras: number;
            /** Incident Counts By Camera */
            incident_counts_by_camera: {
                [key: string]: number;
            };
        };
// ...
        IncidentAnalyticsSchema: {
            /** Total */
            total: number;
            /** Counts By Status */
            counts_by_status: {
                [key: string]: number;
            };
        };
// ...
        IncidentStatus: "NEW" | "ACTIVE" | "ACKNOWLEDGED" | "RESOLVED" | "ARCHIVED";
// ...
        ThreatAnalyticsSchema: {
            /** Counts By Threat Level */
            counts_by_threat_level: {
                [key: string]: number;
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}